Add hospital model + filter by hospital
