+++ app/templates/room_request/hospitals.html
@@ -1,0 +1,14 @@
+{% extends 'layouts/base.html' %}
+{% import 'macros/form_macros.html' as f %}
+
+{% block content %}
+    <div class="ui stackable centered grid container">
+        <div class="twelve wide column">
+            <h2 class="ui header">
+                Edit Hospital Options
+            </h2>
+            
+            {{ f.render_form(form) }}
+        </div>
+    </div>
+{% endblock %}
