+++ app/utils.py
@@ -21,24 +21,3 @@ def register_template_utils(app):
 
 def index_for_role(role):
     return url_for(role.index)
-
-
-class CustomSelectField(Field):
-    widget = HiddenInput()
-
-    def __init__(self, label='', validators=None, multiple=False,
-                 choices=[], allow_custom=True, **kwargs):
-        super(CustomSelectField, self).__init__(label, validators, **kwargs)
-        self.multiple = multiple
-        self.choices = choices
-        self.allow_custom = allow_custom
-
-    def _value(self):
-        return text_type(self.data) if self.data is not None else ''
-
-    def process_formdata(self, valuelist):
-        if valuelist:
-            self.data = valuelist[1]
-            self.raw_data = [valuelist[1]]
-        else:
-            self.data = ''
