+++ manage.py
@@ -8,7 +8,7 @@ from redis import Redis
 from rq import Connection, Queue, Worker
 
 from app import create_app, db
-from app.models import Role, User, Guest, RoomRequest
+from app.models import Role, User, Hospital, RoomRequest
 from config import Config
 
 app = create_app(os.getenv('FLASK_CONFIG') or 'default')
@@ -54,6 +54,7 @@ def add_fake_data(number_users):
     Adds fake data to the database.
     """
     User.generate_fake(count=number_users)
+    Hospital.generate_fake()
     RoomRequest.generate_fake()    
 
 @manager.command
